Improve BBO events in order book reconstruction
Trigger a BBO event when the first order is added to an order book.

Trigger a BBO event when the last bid (or ask) in an order book is
fully executed or canceled, leaving only asks (or bids).

Trigger a BBO event when the only order in an order book is fully
executed or canceled, leaving an empty order book.