Switch from single-channel to multi-channel MoldUDP64 client
The single-channel MoldUDP64 client uses the same channel for
receiving downstream packets and sending request packets. This
results in the source UDP port in request packets being equal
to the UDP port of the IP multicast. Thus, only one instance
of this implementation can be listening to a particular session
on a particular host at a time.

Switch to the multi-channel MoldUDP64 client. It uses separate
channels for receiving downstream packets and sending request
packets.