Adjust receive buffer size in trading system
Set the SoupBinTCP receive buffer size to the maximum size of an inbound
message in the order entry protocol.

Nassau 0.3.0 did not set a lower bound for the SoupBinTCP receive buffer
size. This caused a problem as the maximum size of an inbound message in
the order entry protocol is smaller than the length of the payload in a
Login Request packet in SoupBinTCP. In order to be able to receive Login
Request packets, the trading system could not freely set the SoupBinTCP
receive buffer size.

As Nassau 0.4.0 sets a lower bound for the SoupBinTCP receive buffer
size based on the Login Request packet, the trading system can now set
the SoupBinTCP receive buffer size based on the order entry protocol.