Fix pending cancel status handling in FIX gateway
When an Order Cancel Request or Order Cancel/Replace Request message
is received, the FIX gateway updates ClOrdID(11) and OrigClOrdID(41)
and sends an Execution Report message with an ExecType(150) value of
6 (Pending cancel) or E (Pending replace). Updating ClOrdID(11) and
OrigClOrdID(41) at this point is an error, because fills should be
still reported using the original ClOrdID(11) value until the FIX
gateway sends an Execution Report message with an ExecType(150) value
of 4 (Canceled) or 5 (Replace). This message is triggered by an Order
Canceled message from the trading system.

Thanks to Pekka Enberg for reporting the failure.