Fix pending cancel status handling in FIX gateway
If the FIX gateway receives an Order Executed message indicating that
an order has been executed fully while awaiting for an Order Canceled
message indicating that the order has been canceled in part or fully,
it must generate an Order Cancel Reject message with an CxlRejReason(102)
value of 0 (Too late to cancel). Implement this.