Remove Broken Trade message from POE protocol specification
The Broken Trade messages in the order entry and market data protocols
are inspired by Nasdaq protocols on which these network interfaces are
based.

However, since the beginning of the project, implementing trade breaking
has never actually been on the roadmap. Right now is a convenient time
to clean up the network interfaces as another feature under development,
support for configurable price and quantity representation, impacts the
protocol specifications as well.